chore: reorganized folder structure, removed ts config for node, added graphql schema and configuration
