+++ .graphqlrc.yml
@@ -1,0 +1,2 @@
+schema: ./schema.graphql
+documents: src/**/*.{ts,tsx}
