+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "dev": "vite",
     "build": "tsc -b && vite build",
-    "lint": "eslint .",
+    "lint": "eslint 'src/**/*.{js,ts,jsx,tsx}'",
     "preview": "vite preview"
   },
   "dependencies": {
@@ -18,6 +18,7 @@
     "@mui/material": "^7.0.1",
     "@types/react-router-dom": "^5.3.3",
     "cv-graphql": "^2.13.2",
+    "get-graphql-schema": "^2.1.2",
     "graphql": "^16.10.0",
     "js-cookie": "^3.0.5",
     "react": "^19.0.0",
