+++ src/pages/auth/verify-email/VerifyEmailPage.tsx
