+++ src/shared/hooks/useAuth.tsx
@@ -1,5 +1,5 @@
 import { createContext, useContext } from "react";
-import { AuthContextType } from "../types/auth";
+import { AuthContextType } from "../types/auth.ts";
 
 export const AuthContext = createContext<AuthContextType | undefined>(undefined);
