+++ src/shared/hooks/useLogin.tsx
@@ -4,7 +4,7 @@ import { useLazyQuery } from "@apollo/client";
 import { useNavigate } from "react-router-dom";
 import Cookies from "js-cookie";
 
-import { LOGIN } from "../graphQL/auth/loginQuery";
+import { LOGIN } from "../graphQL/auth/loginQuery.ts";
 
 export const useLogin = () => {
   const navigate = useNavigate();
@@ -15,7 +15,7 @@ export const useLogin = () => {
       const response = await loginQuery({ variables: { auth: input } });
       const accessToken = response?.data?.login?.access_token;
       const user = response?.data?.login?.user;
-  
+
       if (!accessToken || !user) {
         throw new Error("Login failed: no access token or user");
       }
