+++ src/shared/lib/checkAuth.ts
